+++ src/NativeStripeSdk.tsx
@@ -69,7 +69,13 @@ type NativeStripeSdkType = {
   initPaymentSheet(
     params: PaymentSheet.SetupParams
   ): Promise<InitPaymentSheetResult>;
-  presentPaymentSheet(): Promise<PresentPaymentSheetResult>;
+  initPaymentSheetWithOrderTracking(
+    params: PaymentSheet.SetupParams,
+    callback?: () => void
+  ): Promise<InitPaymentSheetResult>;
+  presentPaymentSheet(
+    options: PaymentSheet.PresentOptions
+  ): Promise<PresentPaymentSheetResult>;
   confirmPaymentSheetPayment(): Promise<ConfirmPaymentSheetPaymentResult>;
   createTokenForCVCUpdate(cvc: string): Promise<CreateTokenForCVCUpdateResult>;
   handleURLCallback(url: string): Promise<boolean>;
@@ -127,6 +133,12 @@ type NativeStripeSdkType = {
   ): Promise<
     PlatformPay.ConfirmPaymentResult | PlatformPay.ConfirmSetupIntentResult
   >;
+  configureOrderTracking(
+    orderTypeIdentifier: string,
+    orderIdentifier: string,
+    webServiceUrl: string,
+    authenticationToken: string
+  ): Promise<void>;
 };
 
 const { StripeSdk } = NativeModules;