+++ src/functions.ts
@@ -405,17 +405,28 @@ export const verifyMicrodepositsForSetup = async (
 export const initPaymentSheet = async (
   params: PaymentSheet.SetupParams
 ): Promise<InitPaymentSheetResult> => {
+  let result;
+
+  const setOrderTracking = params?.applePay?.setOrderTracking;
   try {
-    const { paymentOption, error } = await NativeStripeSdk.initPaymentSheet(
-      params
-    );
-    if (error) {
+    if (setOrderTracking && Platform.OS === 'ios') {
+      result = await NativeStripeSdk.initPaymentSheetWithOrderTracking(
+        params,
+        () => {
+          setOrderTracking(NativeStripeSdk.configureOrderTracking);
+        }
+      );
+    } else {
+      result = await NativeStripeSdk.initPaymentSheet(params);
+    }
+
+    if (result.error) {
       return {
-        error,
+        error: result.error,
       };
     }
     return {
-      paymentOption,
+      paymentOption: result.paymentOption,
     };
   } catch (error: any) {
     return {
@@ -424,25 +435,27 @@ export const initPaymentSheet = async (
   }
 };
 
-export const presentPaymentSheet =
-  async (): Promise<PresentPaymentSheetResult> => {
-    try {
-      const { paymentOption, error } =
-        await NativeStripeSdk.presentPaymentSheet();
-      if (error) {
-        return {
-          error,
-        };
-      }
-      return {
-        paymentOption: paymentOption!,
-      };
-    } catch (error: any) {
+export const presentPaymentSheet = async (
+  options: PaymentSheet.PresentOptions = {}
+): Promise<PresentPaymentSheetResult> => {
+  try {
+    const { paymentOption, error } = await NativeStripeSdk.presentPaymentSheet(
+      options
+    );
+    if (error) {
       return {
         error,
       };
     }
-  };
+    return {
+      paymentOption: paymentOption!,
+    };
+  } catch (error: any) {
+    return {
+      error,
+    };
+  }
+};
 
 export const confirmPaymentSheetPayment =
   async (): Promise<ConfirmPaymentSheetPaymentResult> => {